created new branch called dev2. also started to construct my new test
