+++ src/main.css
@@ -4,7 +4,6 @@ body {
     text-align: center;
     font-size: 20px;
     font-family: 'Open Sans', sans-serif;
-
 }
 h1 {
     font-size: 100px;
