+++ test/index.js
@@ -1,11 +1,15 @@
 /* eslint-disable */
 
+// include jsdom for DOM use in tests on travis
 const jsdom = require('jsdom');
 const { JSDOM } = jsdom;
-const { window } = new JSDOM(``);
+const { window } = new JSDOM(``, {
+    url: 'http://localhost:5500'
+});
 global.window = window;
 global.document = window.document;
 global.FormData = window.FormData;
+global.localStorage = window.localStorage;
 
 require = require('esm')(module);
 module.exports = require('./tests.js');
